Remove hardcoded prettier defaults for better compatibility
Some of Prettier's defaults changed in version 2.0, so the defaults in this vscode plugin were updated to match. This caused configuration to break for prettier 1.x projects without explicit config files. This commit adds a new 'default' option to every one of these settings. When this is selected, the plugin does not pass any value for that setting to prettier, allowing the version-specific default to be used.