Fix worker call IDs with multiple worker instances
Previously, 'currentCallMethodId' was a field on each instance of
PrettierWorkerInstance, however, there was only a single instance of 'worker'.
So when there were multiple instances of PrettierWorkerInstance, they would
each have their own 'currentCallMethodId' which could overlap. This meant that
the instances could not reliably get results back from the worker, as there
were multiple results with the same ID. This results in many issues, such as
overwriting files with the contents of other files while formatting.

This change moves 'currentCallMethodId' outside PrettierWorkerInstance, so
that all instances will use the same value and cannot have overlapping IDs.