Use ID in all messages to worker
The same issue that existed for the callMethod messages also existed for the
import messages, where multiple PrettierWorkerInstances could get messages
meant for a different instance, so the 'version' field could end up with the
wrong value.

This changes the code to include an ID in all messages sent to the worker, so
that the PrettierWorkerInstances can properly determine whether a message is
meant for them or not.